Move the apps table to js
- move the table tools to the action menu
- add app tabs
- plug the app tabs in the global search results
- add a app name cell decorator to open the app tabs
- enhance the apps rest api to support the tools

+++ init/models/rest/api_apps.py
@@ -1,3 +1,13 @@
+def lib_app_id(id):
+    try:
+        id = int(id)
+        return id
+    except:
+        pass
+    q = db.apps.app == id
+    row = db(q).select(db.apps.id).first()
+    if row is not None:
+        return row.id
 
 #
 class rest_get_apps(rest_get_table_handler):
@@ -17,7 +27,6 @@ class rest_get_apps(rest_get_table_handler):
           desc=desc,
           examples=examples,
         )
-
 
 #
 class rest_get_app(rest_get_line_handler):
@@ -71,6 +80,11 @@ class rest_post_apps(rest_post_handler):
         )
 
     def handler(self, **vars):
+        if "id" in vars:
+            app_id = vars["id"]
+            del(vars["id"])
+            return rest_post_app(app_id, **vars)
+
         check_privilege("AppManager")
         if len(vars) == 0 or "app" not in vars:
             raise Exception("Insufficient data")
@@ -135,6 +149,35 @@ class rest_post_app(rest_post_handler):
 
 
 #
+class rest_delete_apps(rest_delete_handler):
+    def __init__(self):
+        desc = [
+          "Delete application codes.",
+          "Also deletes all responsible group attachments",
+          "The user must be in the AppManager privilege group.",
+          "The action is logged in the collector's log.",
+          "A websocket event is sent to announce the change in the table.",
+        ]
+        examples = [
+          """# curl -u %(email)s -o- -X DELETE https://%(collector)s/init/rest/api/apps/MYPROJ""",
+        ]
+        rest_delete_handler.__init__(
+          self,
+          path="/apps",
+          desc=desc,
+          examples=examples,
+        )
+
+    def handler(self, **vars):
+        if "id" in vars:
+            app_id = vars["id"]
+        elif "app" in vars:
+            app_id = vars["app"]
+        else:
+            raise Exception("Either the 'id' or 'app' key is mandatory")
+        return rest_delete_app().handler(app_id)
+
+#
 class rest_delete_app(rest_delete_handler):
     def __init__(self):
         desc = [
@@ -159,6 +202,9 @@ class rest_delete_app(rest_delete_handler):
         if id is None:
             return dict(info="app code does not exist")
         q = db.apps.id == id
+        row = db(q).select().first()
+        if row is None:
+            return dict(info="app code %s does not exist" % str(id))
         db(q).delete()
         q = db.apps_responsibles.app_id == row.id
         db(q).delete()
@@ -251,6 +297,28 @@ class rest_get_app_services(rest_get_table_handler):
         self.set_q(q)
         return self.prepare_data(**vars)
 
+class rest_post_apps_responsibles(rest_post_handler):
+    def __init__(self):
+        desc = [
+          "Attach responsible groups to application codes.",
+        ]
+        examples = [
+          "# curl -u %(email)s --header 'Content-Type: application/json' -d @/tmp/data.json -o- -X POST https://%(collector)s/init/rest/api/apps_responsibles",
+        ]
+        rest_post_handler.__init__(
+          self,
+          path="/apps_responsibles",
+          desc=desc,
+          examples=examples
+        )
+
+    def handler(self, **vars):
+        if not "app_id" in vars:
+            raise Exception("The 'app_id' key is mandatory")
+        if not "group_id" in vars:
+            raise Exception("The 'group_id' key is mandatory")
+        return rest_post_app_responsible().handler(vars["app_id"], vars["group_id"])
+
 class rest_post_app_responsible(rest_post_handler):
     def __init__(self):
         desc = [
@@ -271,17 +339,64 @@ class rest_post_app_responsible(rest_post_handler):
         app_id = lib_app_id(app_id)
         if app_id is None:
             raise Exception("app code not found")
+        app = db(db.apps.id==app_id).select().first()
         try:
             group_id = int(group_id)
         except:
             group_id = lib_group_id(group_id)
         if group_id is None:
             raise Exception("group not found")
-        try:
-            attach_group_to_app(group_id, app_id)
-        except CompError as e:
-            return dict(error=str(e))
-        return dict(info="group attached")
+        group = db(db.auth_group.id==group_id).select().first()
+
+        q = db.apps_responsibles.app_id == app_id
+        q &= db.apps_responsibles.group_id == group_id
+        row = db(q).select().first()
+        if row is not None:
+            raise Exception("group %s is already responsible for app %s" % (group.role, app.app))
+
+        id = db.apps_responsibles.insert(app_id=app_id, group_id=group_id)
+
+        fmt = 'attached group %(g)s to app %(u)s'
+        d = dict(g=group.role, u=app.app)
+
+        table_modified("apps_responsibles")
+        _log('apps.group.attach', fmt, d)
+
+        l = {
+          'event': 'apps_change',
+          'data': {'id': id},
+        }
+        _websocket_send(event_msg(l))
+
+        # remove dashboard alerts
+        q = db.dashboard.dash_type == "application code without responsible"
+        q &= db.dashboard.dash_dict.like('%%:"%s"%%'%app.app)
+        db(q).delete()
+        table_modified("dashboard")
+
+        return dict(info=fmt % d)
+
+class rest_delete_apps_responsibles(rest_delete_handler):
+    def __init__(self):
+        desc = [
+          "Detach responsible groups from application codes.",
+        ]
+        examples = [
+          "# curl -u %(email)s -o- --header 'Content-Type: application/json' -d @/tmp/data.json -X DELETE https://%(collector)s/init/rest/api/apps_responsibles",
+        ]
+        rest_delete_handler.__init__(
+          self,
+          path="/apps_responsibles",
+          desc=desc,
+          examples=examples
+        )
+
+    def handler(self, **vars):
+        if not "app_id" in vars:
+            raise Exception("The 'app_id' key is mandatory")
+        if not "group_id" in vars:
+            raise Exception("The 'group_id' key is mandatory")
+        return rest_delete_app_responsible().handler(vars["app_id"], vars["group_id"])
 
 class rest_delete_app_responsible(rest_delete_handler):
     def __init__(self):
@@ -303,17 +418,36 @@ class rest_delete_app_responsible(rest_delete_handler):
         app_id = lib_app_id(app_id)
         if app_id is None:
             raise Exception("app code not found")
+        app = db(db.apps.id==app_id).select().first()
         try:
             group_id = int(group_id)
         except:
             group_id = lib_group_id(group_id)
         if group_id is None:
             raise Exception("group not found")
-        try:
-            detach_group_from_app(group_id, app_id)
-        except CompError as e:
-            return dict(error=str(e))
-        return dict(info="group detached")
+        group = db(db.auth_group.id==group_id).select().first()
+
+        q = db.apps_responsibles.app_id == app_id
+        q &= db.apps_responsibles.group_id == group_id
+        row = db(q).select().first()
+        if row is None:
+            raise Exception("group %s is already not responsible for app %s" % (group.role, app.app))
+
+        db(q).delete()
+
+        fmt = 'detached group %(g)s from app %(u)s'
+        d = dict(g=group.role, u=app.app)
+
+        table_modified("apps_responsibles")
+        _log('apps.group.detach', fmt, d)
+
+        l = {
+          'event': 'apps_change',
+          'data': {'id': row.id},
+        }
+        _websocket_send(event_msg(l))
+
+        return dict(info=fmt % d)
 
 #
 class rest_get_app_responsibles(rest_get_table_handler):
